M*SS:1,修改为空闲状态发送心跳包 2,注册会员单位改为必填
