+++ app/src/main/java/com/xinyu/mwp/activity/RegisterActivity.java
@@ -88,7 +88,7 @@ public class RegisterActivity extends BaseControllerActivity {
 
         setTitle(title);
         phoneEditText.setInputType(EditorInfo.TYPE_CLASS_PHONE);
-        checkHelper.checkButtonState(nextButton, phoneEditText, msgEditText, pwdEditText);
+        checkHelper.checkButtonState(nextButton, phoneEditText, msgEditText, pwdEditText, memberUnit, agentId, refereeId);
         checkHelper.checkVerificationCode(msgEditText.getRightText(), phoneEditText);
         memberUnit.getEditText().setInputType(InputType.TYPE_CLASS_NUMBER);
     }
@@ -106,7 +106,7 @@ public class RegisterActivity extends BaseControllerActivity {
                 }
                 int verifyType = 0;// 0-注册 1-登录 2-更新服务
                 int type = 0;  //0-注册
-                VerifyCodeUtils.getCode(msgEditText, verifyType, context, view, phoneEditText,type);
+                VerifyCodeUtils.getCode(msgEditText, verifyType, context, view, phoneEditText, type);
             }
         });
 
@@ -172,7 +172,7 @@ public class RegisterActivity extends BaseControllerActivity {
     private void register() {
         //本地校验验证码   MD5(yd1742653sd + code_time + rand_code + phone)
         if (!RegisterVerifyCodeEntry.vToken.equals(MD5Util.MD5("yd1742653sd" + RegisterVerifyCodeEntry.timeStamp + vCode))) {
-           ToastUtils.show(context,"验证码错误,请重新输入");
+            ToastUtils.show(context, "验证码错误,请重新输入");
             closeLoader();
             return;
         }
