+++ app/src/main/java/com/xinyu/mwp/application/MyApplication.java
@@ -3,7 +3,6 @@ package com.xinyu.mwp.application;
 import android.app.Activity;
 import android.app.NotificationManager;
 import android.content.Context;
-import android.content.Intent;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.os.Handler;
@@ -23,7 +22,6 @@ import com.xinyu.mwp.entity.EventBusMessage;
 import com.xinyu.mwp.entity.LoginReturnEntity;
 import com.xinyu.mwp.entity.UserEntity;
 import com.xinyu.mwp.listener.OnAPIListener;
-import com.xinyu.mwp.networkapi.HeartPackageService;
 import com.xinyu.mwp.networkapi.Host;
 import com.xinyu.mwp.networkapi.NetworkAPIConfig;
 import com.xinyu.mwp.networkapi.NetworkAPIFactoryImpl;
@@ -33,7 +31,6 @@ import com.xinyu.mwp.user.UserManager;
 import com.xinyu.mwp.util.FileCacheUtil;
 import com.xinyu.mwp.util.LogUtil;
 import com.xinyu.mwp.util.MD5Util;
-import com.xinyu.mwp.util.ServiceUtils;
 import com.xinyu.mwp.util.Utils;
 
 import org.greenrobot.eventbus.EventBus;
@@ -51,7 +48,6 @@ public class MyApplication extends MultiDexApplication implements OnUserUpdateLi
     private static MyApplication application;
     public static List<Activity> activityList = new ArrayList<Activity>();
     public static Handler mainHandler;
-    private Intent intent;
 
     @Override
     public void onCreate() {
@@ -59,7 +55,6 @@ public class MyApplication extends MultiDexApplication implements OnUserUpdateLi
         super.onCreate();
         application = this;
         initNetworkAPIConfig();
-        initHeartPackage();
         initImageLoader();
         initUser();
         checkToken();
@@ -164,8 +159,6 @@ public class MyApplication extends MultiDexApplication implements OnUserUpdateLi
                     .getSystemService(Context.NOTIFICATION_SERVICE))
                     .cancelAll();
         }
-        LogUtil.d("程序退出了----关掉服务---------------");
-        stopService(intent);
         android.os.Process.killProcess(android.os.Process.myPid());
     }
 
@@ -259,14 +252,5 @@ public class MyApplication extends MultiDexApplication implements OnUserUpdateLi
             e.printStackTrace();
             return -1;
         }
-    }
-
-    private void initHeartPackage() {
-        LogUtil.d("服务sss的运行状态是:" + ServiceUtils.isServiceRunning(this, "com.xinyu.mwp.networkapi.HeartPackageService"));
-        intent = new Intent(this, HeartPackageService.class);
-        if (!ServiceUtils.isServiceRunning(this, "com.xinyu.mwp.networkapi.HeartPackageService")) {
-            startService(intent);
-        }
-
     }
 }
