--- app/src/main/java/com/xinyu/mwp/util/ServiceUtils.java
@@ -1,28 +1,0 @@
-package com.xinyu.mwp.util;
-
-import android.app.ActivityManager;
-import android.content.Context;
-
-import java.util.List;
-
-/**
- * Created by Administrator on 2017/5/17.
- */
-public class ServiceUtils {
-    /**
-     * 校验某个服务是否还存在
-     */
-    public static boolean isServiceRunning(Context context, String serviceName) {
-        ActivityManager am = (ActivityManager) context
-                .getSystemService(Context.ACTIVITY_SERVICE);
-        List<ActivityManager.RunningServiceInfo> services = am.getRunningServices(100);
-        for (ActivityManager.RunningServiceInfo info : services) {
-            // 得到所有正在运行的服务的名称
-            String name = info.service.getClassName();
-            if (serviceName.equals(name)) {
-                return true;
-            }
-        }
-        return false;
-    }
-}
